Added Session
added session stuff, and also about page fixed the log in page so it shows the users email and a logout section as well. and a little cart logo to the cart on the main page. didnt move it over to other ones yet not that there are many pages.